+++ dbt/include/snowflake/macros/adapters.sql
@@ -293,3 +293,12 @@
     {{ snowflake_dml_explicit_transaction(truncate_dml) }}
   {%- endcall %}
 {% endmacro %}
+
+{% macro snowflake__copy_grants() %}
+    {% set copy_grants = config.get('copy_grants', False) %}
+    {{ return(copy_grants) }}
+{% endmacro %}
+
+{%- macro snowflake__support_multiple_grantees_per_dcl_statement() -%}
+    {{ return(False) }}
+{%- endmacro -%}